requirements.txt
boto3==1.16.18
botocore==1.19.18
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
ecdsa==0.14.1
envs==1.3
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
pyasn1==0.4.8
pycognito==0.1.4
pycparser==2.20
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
shotgun-api3==3.2.4
six==1.15.0
urllib3==1.26.2
git+git://github.com/shotgunsoftware/python-api.git